Apply patch HARMONY-5557 ([build] Problem building Harmony on MSVC 2005)
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@641826 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/thread/src/thread_native_basic.c
@@ -336,7 +336,7 @@ hythread_t hythread_self_slow() {
 void VMCALL hythread_set_self(hythread_t thread) {
 #ifndef _WIN64
 #   if (_MSC_VER >= 1400)
-        __writefsdword(offsetof(NT_TIB, ArbitraryUserPointer), thread);
+        __writefsdword(offsetof(NT_TIB, ArbitraryUserPointer), (unsigned long)thread);
 #   else
         _asm{
             mov eax, thread
